refactor(import): extract CustomerUpserter shared service (ADR-026 P2/P3)
Move the idempotent Tempo->Customer upsert out of
ProjectImportConfirmationService into a CustomerUpserter so the P1 confirm
flow and the P3 import auto-create share one upsert rather than duplicating
the logic. Behaviour-preserving: the confirm service delegates to it.
